move executeUpgradeTransaction into util.js so that it can be used by core.js without bringing in all the upgrade tests
